Add test, fix bug
